refactor: 将应用程序中的用户和文件 ID 类型从 Integer 更改为 Long
- 更新了各个实体、DTO 和服务接口中的 ID 字段，使其使用 Long 类型而非 Integer 类型，以提高可扩展性。

- 调整了相关方法和数据库模式，以反映 ID 类型的更改。

- 确保控制器和服务中用户 ID 处理的一致性。

+++ aether-admin/src/main/java/com/xiaobai1226/aether/admin/domain/vo/UpdateUserVO.java
@@ -20,7 +20,7 @@ public class UpdateUserVO {
      * 用户ID
      */
     @NotNull(message = ERROR_USER_ID_EMPTY)
-    private Integer id;
+    private Long id;
 
     /**
      * 用户名
@@ -42,7 +42,7 @@ public class UpdateUserVO {
     /**
      * 密码
      */
-    private Integer roleId;
+    private Long roleId;
 
     /**
      * 用户状态
